+++ backend_mercos/models/pedido.py
@@ -8,6 +8,6 @@ class Pedido(models.Model):
     usuario = models.ForeignKey(Usuario, on_delete=models.CASCADE, default=None)
     cliente = models.ForeignKey(Cliente, on_delete=models.CASCADE, default=None)
     quantidadeItem = models.BigIntegerField(default=0)
-    valor = models.BigIntegerField(default=0)
+    valor = models.FloatField(default=0)
     rentabilidade = models.CharField(max_length=60,
     choices = [(tag, tag.value) for tag in TipoRentabilidade])
